move proofs api to app router
