+++ apps/web/src/utils/proofs/sybil_resistance.ts
@@ -1,6 +1,6 @@
 import { getAttestations } from '@coinbase/onchainkit/identity';
 import { kv } from 'apps/web/src/utils/datastores/kv';
-import { CoinbaseProofResponse } from 'apps/web/pages/api/proofs/coinbase';
+import { CoinbaseProofResponse } from 'apps/web/app/(basenames)/api/proofs/coinbase/route';
 import RegistrarControllerABI from 'apps/web/src/abis/RegistrarControllerABI';
 import {
   USERNAME_CB1_DISCOUNT_VALIDATORS,
